fix: extract computeQualitySummary to its own module to break transitive runner.ts dependency
quality-summary.test.ts imported computeQualitySummary from orchestrator/index.ts,
which transitively loads runner.ts and attempts to resolve the claude CLI at module
load time — failing in CI where claude is not installed.

Extract the pure helper to src/application/orchestrator/quality.ts (no infra imports),
re-export it from index.ts for backwards compatibility, and update the test and MCP
server to import directly from the new module.
